+++ src/main/java/roomescape/controller/RoomEscapeController.java
@@ -2,39 +2,26 @@ package roomescape.controller;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.repository.ReservationRepository;
 
-@Controller
+@RestController
 public class RoomEscapeController {
 
     private final ReservationRepository repository = new ReservationRepository();
 
-    @GetMapping("/admin")
-    public String adminPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String reservationAdminPage() {
-        return "admin/reservation-legacy";
-    }
-
-    @ResponseBody
     @GetMapping("/reservations")
     public List<ReservationResponse> reservationsJson() {
         return ReservationResponse.reservationsToDtos(repository.findAll());
     }
 
-    @ResponseBody
     @PostMapping("/reservations")
     public ReservationResponse addReservation(@RequestBody ReservationRequest request) {
         Long id = repository.add(request.dtoToReservationWithoutId());
@@ -43,7 +30,11 @@ public class RoomEscapeController {
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
-        repository.deleteById(id);
-        return ResponseEntity.ok().build();
+        try {
+            repository.deleteById(id);
+            return ResponseEntity.ok().build();
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().build();
+        }
     }
 }
